requirements.txt
torch
torchvision
timm
opencv-python
git+https://github.com/dhkim2810/MobileSAM.git
matplotlib
fastapi
uvicorn
pillow
numpy
